[REFACTOR] ClothView ClothViewModel
