+++ POMPOM/POMPOM/UI/ClothPickerView/ClothPickerViewModel.swift
@@ -36,8 +36,6 @@ final class PickerCombineViewModel: ObservableObject {
             }
             .store(in: &cancellables)
         
-        
-        
         colorSubject
             .compactMap { $0 }
             .removeDuplicates()
@@ -162,7 +160,7 @@ class PickerViewModel: ClothesViewModel {
     func selectItem(name: String, hex: String) {
         if selectedItems[currentType] != nil {
             if selectedItems[currentType]?.id == name {
-                selectedItems[currentType] = Cloth(id: " ", hex: " ", category: currentType)
+                selectedItems[currentType] = Cloth(id: " ", hex: hex, category: currentType)
             } else {
                 //해당 key 에 해당하는 객체가 이미 존재하는 경우에는 새 객체를 생성하는 것이 아닌 값만 변경해준다.
                 selectedItems[currentType]?.hex = hex
